Use .container/.component naming pattern for ens-input
