+++ ui/app/pages/send/send-content/add-recipient/ens-input.container.js
@@ -1,0 +1,23 @@
+import EnsInput from './ens-input.component'
+import {getAddressBook, getCurrentNetwork, getSendTo, getSendToNickname} from '../../send.selectors'
+import {
+  updateSendTo,
+  updateEnsResolution,
+  updateEnsResolutionError,
+} from '../../../../store/actions'
+const connect = require('react-redux').connect
+
+
+export default connect(
+  state => ({
+    network: getCurrentNetwork(state),
+    selectedAddress: getSendTo(state),
+    selectedName: getSendToNickname(state),
+    addressBook: getAddressBook(state),
+  }),
+  dispatch => ({
+    updateSendTo: (to, nickname) => dispatch(updateSendTo(to, nickname)),
+    updateEnsResolution: (ensResolution) => dispatch(updateEnsResolution(ensResolution)),
+    updateEnsResolutionError: (message) => dispatch(updateEnsResolutionError(message)),
+  })
+)(EnsInput)
